feat: 훅과 api로 분리
