+++ src/components/templates/news/news/News.tsx
@@ -6,36 +6,43 @@ import WriteButton from "~/components/atoms/news/news/WriteButton";
 import Pagination from "../../../molecules/news/Pagination";
 import Search from "~/components/molecules/news/Search";
 import { handleSearch } from "~/hooks/handleSearch";
-import { useNewsData } from "~/hooks/board/useBoardData";
+import { useGetBoards } from "~/hooks/board/useBoardData";
 
 const News = () => {
-  const { data: boardlist = [], isLoading, isError } = useNewsData();
-  const [filteredBoardlist, setFilteredBoardlist] = useState(boardlist);
+  const { data: newsData = [], isLoading, error } = useGetBoards("news");
+  const [filteredBoardlist, setFilteredBoardlist] = useState(newsData);
   const [page, setPage] = useState(1);
   const [searchQuery, setSearchQuery] = useState("");
   const [searchFilterOption, setsearchFilterOption] = useState("title");
   const limit = 15;
   const offset = (page - 1) * limit;
 
-  console.log(boardlist);
-
   useEffect(() => {
-    setFilteredBoardlist(boardlist);
-  }, [boardlist]);
+    setFilteredBoardlist(newsData);
+  }, [newsData]);
 
   const showSearch = () => {
-    const result = handleSearch(boardlist, searchQuery, searchFilterOption);
+    const result = handleSearch(newsData, searchQuery, searchFilterOption);
     setFilteredBoardlist(result);
+
     setPage(1);
   };
 
   if (isLoading) {
-    return <div>Loading...</div>;
+    return <div>로딩중입니다</div>;
   }
 
-  if (isError) {
-    return;
+  if (error) {
+    return <div>에러</div>;
   }
+
+  // if (isLoading) {
+  //   return <PendingMessage />;
+  // }
+
+  // if (error) {
+  //   return <ErrorMessage text="에러" />;
+  // }
 
   return (
     <NewsLayout>
